third_party/WebKit/LayoutTests/fullscreen/full-screen-no-style-sharing.html

<body>
    <video id="video1" width="1"></video>
    <video id="video2" width="1"></video>
    <script src="full-screen-test.js"></script>
    <script src=../media/media-file.js></script>
    <script>
        var video2 = document.getElementById('video2');
        waitForEventTestAndEnd(document, 'webkitfullscreenchange', "video2.clientWidth==document.body.clientWidth");
        runWithKeyDown(function(){video2.webkitRequestFullScreen()});
    </script>
</body>
