third_party/blink/web_tests/fullscreen/full-screen-no-style-sharing-expected.txt

EVENT(webkitfullscreenchange) TEST(video2.clientWidth==document.body.clientWidth) OK
END OF TEST
